refactor: db folder structure
