+++ apps/studio/database/database.ts
@@ -1,5 +1,5 @@
 import { Kysely, PostgresDialect } from 'kysely'
-import { DB } from '../../prisma/generated/types'
+import { DB } from './prisma/generated/types'
 import { Pool } from 'pg'
 
 const connectionString = `${process.env.DATABASE_URL}`
@@ -10,6 +10,6 @@ const dialect = new PostgresDialect({
   }),
 })
 
-export const db = new Kysely<DB>({
+export const db: Kysely<DB> = new Kysely<DB>({
   dialect,
 })
