+++ apps/studio/database/prisma/seed.ts
@@ -1,0 +1,19 @@
+/**
+ * Adds seed data to your db
+ *
+ * @link https://www.prisma.io/docs/guides/database/seed-database
+ */
+import { db } from '../database'
+
+async function main() {
+  // Nothing
+}
+
+main()
+  .catch((e) => {
+    console.error(e)
+    process.exit(1)
+  })
+  .finally(async () => {
+    await db.destroy()
+  })
