--- apps/studio/prisma/seed.ts
@@ -1,15 +1,0 @@
-/**
- * Adds seed data to your db
- *
- * @link https://www.prisma.io/docs/guides/database/seed-database
- */
-import { db } from '~/utils/database'
-
-async function main() {
-  // Nothing
-}
-
-main().catch((e) => {
-  console.error(e)
-  process.exit(1)
-})
