+++ apps/studio/src/server/context.ts
@@ -2,7 +2,7 @@ import type * as trpc from '@trpc/server'
 import type { CreateNextContextOptions } from '@trpc/server/adapters/next'
 import { getIronSession } from 'iron-session'
 import { prisma } from './prisma'
-import { db } from '../utils/database'
+import { db } from '../../database/database'
 import { sessionOptions } from './modules/auth/session'
 import { type SessionData, type Session } from '~/lib/types/session'
 import { type User } from '@prisma/client'
